+++ CHANGELOG.md
@@ -22,12 +22,27 @@ any parts of the framework not mentioned in the documentation should generally b
 * Replaced `OrderedDict` with `dict` which is also ordered since Python 3.7.
 * Compound document "include" parameter is only included in the OpenAPI schema if serializer
   implements `included_serializers`.
+* Allowed overwriting of resource id by defining an `id` field on the serializer.
+
+Example:
+```python
+class CustomIdSerializer(serializers.Serializer):
+    id = serializers.CharField(source='name')
+    body = serializers.CharField()
+```
+* Allowed overwriting resource id on resource related fields by creating custom `ResourceRelatedField`.
+
+Example:
+```python
+class CustomResourceRelatedField(relations.ResourceRelatedField):
+    def get_resource_id(self, value):
+        return value.name
+```
 
 ### Fixed
 
 * Refactored handling of the `sort` query parameter to fix duplicate declaration in the generated schema definition
 * Non-field serializer errors are given a source.pointer value of "/data".
-* Serialization of non-model `Serializer` results, e.g. `dict` without a `pk` attribute
 * Fixed "id" field being added to /data/attributes in the OpenAPI schema when it is not rendered there.
 
 ## [6.0.0] - 2022-09-24